Use `SysPathPrefixOrigin` instead of `RangedValue`
