+++ crates/ty_test/src/lib.rs
@@ -12,7 +12,6 @@ use ruff_db::diagnostic::{
 use ruff_db::files::{File, system_path_to_file};
 use ruff_db::panic::catch_unwind;
 use ruff_db::parsed::parsed_module;
-use ruff_db::ranged_value::RangedValue;
 use ruff_db::system::{DbWithWritableSystem as _, SystemPath, SystemPathBuf};
 use ruff_db::testing::{setup_logging, setup_logging_with_filter};
 use ruff_db::{Db as _, Upcast};
@@ -23,7 +22,7 @@ use ty_python_semantic::pull_types::pull_types;
 use ty_python_semantic::types::check_types;
 use ty_python_semantic::{
     Program, ProgramSettings, PythonPath, PythonPlatform, PythonVersionSource,
-    PythonVersionWithSource, SearchPathSettings,
+    PythonVersionWithSource, SearchPathSettings, SysPrefixPathOrigin,
 };
 
 mod assertion;
@@ -275,7 +274,10 @@ fn run_test(
             python_path: configuration
                 .python()
                 .map(|sys_prefix| {
-                    PythonPath::IntoSysPrefix(RangedValue::cli(sys_prefix.to_path_buf()))
+                    PythonPath::sys_prefix(
+                        sys_prefix.to_path_buf(),
+                        SysPrefixPathOrigin::PythonCliFlag,
+                    )
                 })
                 .unwrap_or(PythonPath::KnownSitePackages(vec![])),
         }
